chore(.kodiak.toml): add .kodiak.toml configuration file for Kodiak CI
The .kodiak.toml file is added to the repository to configure the
behavior of the Kodiak CI tool. This file specifies the version of the
configuration file, sets up automatic merging of pull requests opened by
the "dependabot" user for "minor" or "patch" version upgrades, and
ignores "major" version upgrades. It also specifies that the
"dependabot" user should be ignored when updating and closing stale
dependency upgrades.