feat(Dockerfile.resticpg): add Dockerfile for resticpg image
A new Dockerfile named Dockerfile.resticpg has been added to the
build/package directory. This Dockerfile is used to build a Docker image
based on Alpine Linux version 3.18.2. The image includes the restic and
postgresql14-client packages installed using the apk package manager.
This Dockerfile is intended to be used for creating a containerized
environment for running resticpg, a tool for backing up and restoring
PostgreSQL databases using restic.