feat(build-resticpg.yaml): add GitHub Actions workflow to build and push resticpg Docker image
The new file `.github/workflows/build-resticpg.yaml` contains a GitHub
Actions workflow that is triggered on pushes to the `develop`, `master`,
and `feat/**` branches. The workflow sets up the environment, caches
Docker layers, logs into DockerHub, retrieves the short SHA of the
commit, and then builds and pushes the resticpg Docker image with
appropriate tags. This workflow enables automated building and pushing
of the resticpg Docker image whenever changes are made to the specified
branches.